+++ brikar-client/src/main/java/com/truward/brikar/client/interceptor/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package groups subpackages that contain entities and helpers for building service clients.
+ *
+ * @author Alexander Shabanov
+ */
+package com.truward.brikar.client.interceptor;