+++ brikar-client/src/main/java/com/truward/brikar/client/rest/RestOperationsProvider.java
@@ -1,0 +1,7 @@
+package com.truward.brikar.client.rest;
+
+/**
+ * @author Alexander Shabanov
+ */
+public final class RestOperationsProvider {
+}