+++ brikar-client/src/test/java/com/truward/brikar/client/test/rest/RestClientBuilderFactoryConnectionTest.java
@@ -2,7 +2,7 @@ package com.truward.brikar.client.test.rest;
 
 import com.truward.brikar.client.binder.RestServiceBinder;
 import com.truward.brikar.client.binder.RestServiceBinderFactory;
-import com.truward.brikar.client.rest.support.StandardRestBinder;
+import com.truward.brikar.client.rest.support.StandardRestClientBuilderFactory;
 import com.truward.it.httpserver.producer.AbstractResponseProducer;
 import com.truward.it.httpserver.support.DefaultItHttpServer;
 import org.apache.http.HttpRequest;
@@ -24,11 +24,11 @@ import java.net.URI;
 import static org.junit.Assert.assertEquals;
 
 /**
- * Integration tests for {@link com.truward.brikar.client.rest.support.StandardRestBinder}.
+ * Integration tests for {@link StandardRestClientBuilderFactory}.
  *
  * @author Alexander Shabanov
  */
-public class RestBinderConnectionTest {
+public class RestClientBuilderFactoryConnectionTest {
   private DefaultItHttpServer httpServer;
   private ThreadPoolTaskExecutor taskExecutor;
 
@@ -39,7 +39,7 @@ public class RestBinderConnectionTest {
     taskExecutor = new ThreadPoolTaskExecutor();
     taskExecutor.setCorePoolSize(2);
     taskExecutor.setKeepAliveSeconds(1);
-    taskExecutor.setBeanName("RestBinderConnectionTest.taskExecutor");
+    taskExecutor.setBeanName("RestClientBuilderFactoryConnectionTest.taskExecutor");
     taskExecutor.afterPropertiesSet();
 
     httpServer = new DefaultItHttpServer(new ExecutorServiceAdapter(taskExecutor));
@@ -60,7 +60,7 @@ public class RestBinderConnectionTest {
 
   @Test
   public void shouldMakeRequest() {
-    try (final StandardRestBinder restBinder = new StandardRestBinder(new StringHttpMessageConverter())) {
+    try (final StandardRestClientBuilderFactory restBinder = new StandardRestClientBuilderFactory(new StringHttpMessageConverter())) {
       restBinder.setRestServiceBinderFactory(new TestRestServiceBinderFactory());
       restBinder.afterPropertiesSet();